+++ src/app/shared/login/login.component.ts
@@ -1,0 +1,27 @@
+import { Component } from '@angular/core';
+import { Router } from '@angular/router';
+import { UserService } from '../user/user.service';
+
+@Component({
+  moduleId: module.id,
+  selector: 'login',
+  templateUrl: 'login.component.html',
+  styleUrls: ['login.component.css']
+})
+
+export class LoginComponent {
+  constructor(
+    private userService: UserService,
+    private router: Router
+  ) {}
+
+  //  listen to the result of the login and after a success redirect the user to the home page
+  onSubmit(email, password) {
+    // this.userService.login()
+    // .subscribe((result) => {
+    //   if (result) {
+    //     this.router.navigate(['']);
+    //   }
+    // });
+  }
+}