+++ src/app/shared/toolbar/toolbar.component.ts
@@ -1,5 +1,8 @@
-import {Component} from '@angular/core';
-import {MdToolbar} from '@angular2-material/toolbar';
+import { Component } from '@angular/core';
+import { MdToolbar } from '@angular2-material/toolbar';
+import { ROUTER_DIRECTIVES } from '@angular/router';
+import { MdIcon, MdIconRegistry } from '@angular2-material/icon';
+import { UserService } from '../user/user.service';
 
 /**
  * This class represents the toolbar component.
@@ -9,6 +12,12 @@ import {MdToolbar} from '@angular2-material/toolbar';
   selector: 'sd-toolbar',
   templateUrl: 'toolbar.component.html',
   styleUrls: ['toolbar.component.css'],
-  directives: [MdToolbar]
+  directives: [ROUTER_DIRECTIVES, MdToolbar, MdIcon],
+  providers: [MdIconRegistry]
 })
-export class ToolbarComponent {}
+export class ToolbarComponent {
+
+  constructor(
+    private userService: UserService
+  ){}
+}